% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncv_calculation.R
\name{calculate_ncv_score}
\alias{calculate_ncv_score}
\title{Use an NCV template to calculate a NCV score for sample of interest}
\usage{
calculate_ncv_score(nipt_sample, ncv_template)
}
\arguments{
\item{nipt_sample}{nipt_sample object of interest}

\item{ncv_template}{ncv_template object, result from \code{\link{prepare_ncv}}}
}
\value{
ncv_result object
}
\description{
Use an NCV template to calculate a NCV score for sample of interest
}
\details{
\code{\link{prepare_ncv}}
}
\examples{
\dontrun{
##Use NCVTemplate to get NCV scores for the sample of interest
ncv_score_13 <- calculate_ncv_score(nipt_sample = sample_of_interest, 
                                    ncv_template = new_ncv_template_13)
}

}
\references{
\href{http://www.ncbi.nlm.nih.gov/pubmed/21519036}{Sehnert et al.}
}

