% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general_functions_beta.R
\name{chrfractions}
\alias{chrfractions}
\title{Calculate chromosomal fraction}
\usage{
chrfractions(nipt_sample)
}
\arguments{
\item{nipt_sample}{NIPTSample to retrieve chromosomal fraction for}
}
\description{
Calculate chromosomal fraction
}

