% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{getcontrolchromosomes}
\alias{getcontrolchromosomes}
\title{Get control chromosomes names}
\usage{
getcontrolchromosomes(nipt_sample, control_chromosomes = control_chromosomes)
}
\arguments{
\item{nipt_sample}{A sample to check wether combined or separated strands 
are used}

\item{control_chromosomes}{Vector with control chromosomes}
}
\description{
Get control chromosomes names
}

