% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{getreadscontrolgroup}
\alias{getreadscontrolgroup}
\title{Get reads per chromosome per control group sample}
\usage{
getreadscontrolgroup(nipt_control_group)
}
\arguments{
\item{nipt_control_group}{The control group to retrieve reads for}
}
\description{
Get reads per chromosome per control group sample
}

