% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_group.R
\name{remove_duplicates_controlgroup}
\alias{remove_duplicates_controlgroup}
\title{Remove duplicate samples from control group}
\usage{
remove_duplicates_controlgroup(nipt_control_group)
}
\arguments{
\item{nipt_control_group}{NIPTControlGroup object}
}
\value{
NIPTControlGroup object
}
\description{
Removes all duplicate samples in control group by samplename.
}
\details{
This functions removes duplicate samples from the control group based on name. 
It returns a new NIPTControlGroup object.
}
\examples{
\dontrun{
new_control_group <- remove_duplicates_controlgroup(nipt_control_group = old_control_group)
}

}

