% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_group.R
\name{remove_sample_controlgroup}
\alias{remove_sample_controlgroup}
\title{Remove a sample by samplename from control group}
\usage{
remove_sample_controlgroup(samplename, nipt_control_group)
}
\arguments{
\item{samplename}{Regular expression string. All matching samplenames 
are removed from the control group}

\item{nipt_control_group}{NIPTControlGroup object to remove samples from}
}
\value{
NIPTControlGroup object
}
\description{
Remove a sample by samplename from control group
}
\details{
This function removes a sample from the `NIPTControlGroup` object by name.
Note that this function uses a regular expression, and if more sample_names 
satisfy the regular expression, they will also be removed. It returns a new 
NIPTControlGroup object.
}

