% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{retrieve_fractions_of_interest}
\alias{retrieve_fractions_of_interest}
\title{Retrieve the chromosomal fractions of a chromosome of interest}
\usage{
retrieve_fractions_of_interest(nipt_sample, chromo_focus, chromosomal_fracs)
}
\arguments{
\item{nipt_sample}{NIPTSample to check wether the strands are combined or 
separated}

\item{chromo_focus}{The chromosome of interest}

\item{chromosomal_fracs}{The chromosomal fractions to extract the 
chromosome of interest from}
}
\description{
Retrieve the chromosomal fractions of a chromosome of interest
}

