% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{getfractionscontrolgroup}
\alias{getfractionscontrolgroup}
\title{Get all chromosomal fractions of a control group}
\usage{
getfractionscontrolgroup(nipt_control_group)
}
\arguments{
\item{nipt_control_group}{The NIPTControlGroup to retrieve the 
chromosomal fraction of every autosome for}
}
\description{
Get all chromosomal fractions of a control group
}

